pyproject.toml
[project]
name = "password_manager"
version = "0.0.1"
requires-python = ">=3.8"
dependencies = [
    "pycryptodome>=3.20.0",
]

[project.scripts]
password_manager = "driver.__main__:main"
